Add test cases for RandomSource. Change rownum to LONG

+++ src/main/java/me/ericfu/lynx/source/random/RandomGenerator.java
@@ -7,10 +7,10 @@ public interface RandomGenerator {
     /**
      * Generate one value
      *
-     * @param rownum an increasing int number starting from 1
+     * @param rownum an increasing LONG number starting from 1
      * @param rand   the Random object
      * @return generated value
      */
-    Object generate(int rownum, Random rand);
+    Object generate(long rownum, Random rand);
 
 }
